web4, analytics, playwright tests (#43)
## Playwright test fixes

This PR fixes issues with tests after changing the target DAO contract
and the latest UI changes.


## Rust web4 contract

There's a web4 contract written in Rust. It's based on the cargo-near
template. A deployed preview can be found here:

https://treasuryrust.testnet.page/

For the Rust contract, there is a build.rs script to insert API keys,
but currently environment variables are not passed to it unless
building/deploying with the `--no-docker` option for cargo-near.

## Local development gateway

There is also a local development gateway setup, that you can also use
to test the web4 page locally. Check out the updated
[README.md](./README.md) for more info.

---------

Co-authored-by: Megha-Dev-19 <100185149+Megha-Dev-19@users.noreply.github.com>