settings page for voting duration (#121)
https://github.com/user-attachments/assets/af591569-480e-4e9b-aa9e-16304380e33b

Lowering voting duration should warn and show proposals that will expire
because of the change


https://github.com/user-attachments/assets/d5829905-0643-4f57-96fe-b4abebe52c60

When a voting duration change proposal is submitted, a toast will show:


https://github.com/user-attachments/assets/1f004b39-d66c-4b1d-aa4a-57d5ed2047b2

The last test here uses a RPC sandbox, interacting with a real instance
of the smart contract. Not on the mainnet, but in a near-workspaces-rs
sandbox. This new sandbox facilitates real interaction with the smart
contracts in tests, without having to do real mainnet transactions. Also
there is a "playground" script which you can use to check how a contract
works. Currently only the sputnik-dao contract is deployed.



resolves #116
resolves #127