rpc sandbox supporting treasury factory (#194)
This PR adds the `treasury-factory.near` contract and the dependencies
`social.near` and `near` contracts to the RPC sandbox which can be used
for Playwright tests and for experimenting in the
[playground](https://github.com/NEAR-DevHub/neardevhub-treasury-dashboard/blob/feat/sandbox-treasury-factory/sandboxrpc/README.md).

The file
[sandboxrpc/playground/treasuryfactory.js](https://github.com/NEAR-DevHub/neardevhub-treasury-dashboard/blob/feat/sandbox-treasury-factory/sandboxrpc/playground/treasuryfactory.js)
shows how to use the treasury factory with the sandbox. You can run this
file with nodejs and test various ways of interaction.

Also there is a [Playwright
test](https://github.com/NEAR-DevHub/neardevhub-treasury-dashboard/blob/feat/sandbox-treasury-factory/playwright-tests/tests/treasury-factory/treasury-factory.spec.js)
that shows how to use it in Playwright and make the needed assertions
that a treasury instance was successfully created. This can be used as a
starting point for creating the full E2E test in
https://github.com/NEAR-DevHub/neardevhub-treasury-dashboard/pull/191

Broken test fixes:

- "Should update existing member permissions" ( the locator for finding
the delete button was not specific enough ).
- "Should whitelist staking pool and create stake delegation request" (
also made the locator more specific ).

resolves #193