Add ledger to web4 wallet selector (#255)
Adds ledger to the wallet selector for web4 pages.


![image](https://github.com/user-attachments/assets/3486b17f-0309-4fa2-8eb5-129ac342e3db)

Also introducing a new RPC proxy that does not try real RPC first and
then modify with local content, but rather looks for the widget locally
first and if not found, then forwarding to real RPC. This will reduce
stress on RPCs significantly, and also then avoid failing tests because
of exceeded rate limits.

fixes #254 
fixes #290