fix: Social DB updates even if sputnik-dao creation fails (#343)
The original idea was to move sputnik dao creation in the end, after the
social db updates, but after a lot of failed attempts on reordering it
seems that sputnik-dao creation requires more gas available than it
uses. Instead the social db update is executed also if the sputnik-dao
creation failed.

fixes #341