Fix 404 page error for lockup (#399)
We can't directly modify the `app.jsx` file for instances created
through the self-flow, as access remains with the creator. While we do
have access via one of our keys, updating it for all instances would be
challenging.

As a quick fix, I’ve made a change in the Navbar—when the page is
`lockup`, it renders the component instead of the passed children.

@petersalomonsen, just flagging this as part of the system upgrade.
Ideally, users should update their widgets to support the new tabs.