System updates for specified instances and developer docs (#480)
Instead of a feature flag as described in #479, this PR adds the
possibility of specifying which instances that will receive a system
update. After some thought I believe this is more usable than
introducing a feature flag that we later will remove. This approach also
means that we can control which instances to upgrade centrally, rather
than having to adjust the configuration of each instance. It might also
be usable for future upgrades if we want to target specific instances
first.

This PR also contains a documentation file, that describes how the
system updates work and how to publish updates.

To publish an update to specified instances, you use the `instances`
property of the update object. If the `instances` property is not
specified, then all instances will be targeted.

Here's an example:

```javascript
 {
        id: 1,
        createdDate: "2025-03-25",
        version: "n/a",
        type: "DAO contract",
        summary: "Update to latest sputnik-dao contract",
        details: "",
        instances: ["infinex.near", "treasury-testing.near"],
        votingRequired: true,
      }
```

Here's how it appears from the different test cases:


https://github.com/user-attachments/assets/24f10614-6ea1-452b-8838-fe8535454429

The documentation here may also contain info that is relevant for users
as requested in #477

Done some refactorings to reduce the number of duplicate `beforeEach`,
`afterEach` by having CDN caching and unrouting attached to the extended
test function.

Refactored the `redirectWeb4` function to be common for sandbox and
non-sandbox testing.

Finally there's also a script here for merging multiple playwright
videos, and annotating with subtitles from the test title ( which was
used to generate the test video here ).

resolves #479
resolves #397

---------

Co-authored-by: Megha <100185149+Megha-Dev-19@users.noreply.github.com>