Move from `workflow_run` to `workflow_call` (#506)
Initially, the workflow was triggered using:

```yaml
workflow_run:
  workflows: ["Prepare Dependencies"]
  types:
    - completed
```

However, this approach only runs on the default branch (`main`) and
doesn't capture changes from the current branch.

To fix this, I switched to using `workflow_call`. This way, the
`test-pipeline` workflow can be triggered explicitly after dependencies
and sandbox artifacts are uploaded, ensuring it runs on the current
branch and includes the latest changes.