adjust to cargo-near-build 0.15.0 (#508)
This one was a hard one to figure out. Github actions runners were
updated to 1.87, and cargo-near-build was updated yesterday. I thought
for long it was because of the Rust upgrade, but then I also discovered
the cargo-near-build update.

- cargo-near-build 0.15.0 does not need the signer-public-id parameter
- adjust to changes in https://github.com/near/near-cli-rs/pull/468
which was released with cargo-near-build 0.15.0

see the successful deployment here:
https://github.com/NEAR-DevHub/neardevhub-treasury-dashboard/actions/runs/15076851187/job/42386136942

fixes #509